fix translations & cleanup
