+++ apps/web/lib/workspace/constants.ts
@@ -1,0 +1,1 @@
+export const DEFAULT_WORKSPACE_NAME = "My workspace";
