+++ apps/web/locales/ja-JP.json
@@ -316,7 +316,6 @@
     "more_options": "その他のオプション",
     "move_down": "下に移動",
     "move_up": "上に移動",
-    "my_product": "マイプロダクト",
     "name": "名前",
     "new_version_available": "Formbricks {version} が利用可能です。今すぐアップグレード！",
     "next": "次へ",
@@ -833,81 +832,12 @@
     "csat": "顧客満足度(CSAT)",
     "csat_description": "顧客満足度スコアを測定します(1-5)",
     "csat_lower_label": "非常に不満",
-    "csat_name": "顧客満足度（CSAT）",
-    "csat_question_10_headline": "他に何かコメント、質問、懸念はありますか？",
-    "csat_question_10_placeholder": "ここに回答を入力してください...",
-    "csat_question_1_headline": "この$[workspaceName]を友人や同僚に勧める可能性はどのくらいありますか？",
-    "csat_question_1_lower_label": "可能性が低い",
-    "csat_question_1_upper_label": "可能性が非常に高い",
-    "csat_question_2_choice_1": "やや満足",
-    "csat_question_2_choice_2": "非常に満足",
-    "csat_question_2_choice_3": "満足も不満もない",
-    "csat_question_2_choice_4": "やや不満",
-    "csat_question_2_choice_5": "非常に不満",
-    "csat_question_2_headline": "全体的に、当社の$[workspaceName]にどの程度満足していますか、または不満がありますか？",
-    "csat_question_2_subheader": "一つ選択してください：",
-    "csat_question_3_choice_1": "非効率的",
-    "csat_question_3_choice_10": "ユニーク",
-    "csat_question_3_choice_2": "便利",
-    "csat_question_3_choice_3": "非実用的",
-    "csat_question_3_choice_4": "高価",
-    "csat_question_3_choice_5": "高品質",
-    "csat_question_3_choice_6": "信頼できる",
-    "csat_question_3_choice_7": "費用対効果が高い",
-    "csat_question_3_choice_8": "低品質",
-    "csat_question_3_choice_9": "信頼できない",
-    "csat_question_3_headline": "当社の$[workspaceName]を説明するのに、次のうちどの言葉を使いますか？",
-    "csat_question_3_subheader": "当てはまるものをすべて選択してください：",
-    "csat_question_4_choice_1": "非常にうまく",
-    "csat_question_4_choice_2": "とてもうまく",
-    "csat_question_4_choice_3": "ある程度うまく",
-    "csat_question_4_choice_4": "あまりうまく",
-    "csat_question_4_choice_5": "全くうまく",
-    "csat_question_4_headline": "当社の$[workspaceName]は、あなたのニーズをどの程度満たしていますか？",
-    "csat_question_4_subheader": "一つ選択してください：",
-    "csat_question_5_choice_1": "非常に高品質",
-    "csat_question_5_choice_2": "高品質",
-    "csat_question_5_choice_3": "低品質",
-    "csat_question_5_choice_4": "非常に低品質",
-    "csat_question_5_choice_5": "どちらでもない",
-    "csat_question_5_headline": "$[workspaceName]の品質をどのように評価しますか？",
-    "csat_question_5_subheader": "一つ選択してください：",
-    "csat_question_6_choice_1": "素晴らしい",
-    "csat_question_6_choice_2": "平均以上",
-    "csat_question_6_choice_3": "平均的",
-    "csat_question_6_choice_4": "平均以下",
-    "csat_question_6_choice_5": "悪い",
-    "csat_question_6_headline": "$[workspaceName]のコストパフォーマンスをどのように評価しますか？",
-    "csat_question_6_subheader": "一つ選択してください：",
-    "csat_question_7_choice_1": "非常に迅速",
-    "csat_question_7_choice_2": "とても迅速",
-    "csat_question_7_choice_3": "ある程度迅速",
-    "csat_question_7_choice_4": "あまり迅速でない",
-    "csat_question_7_choice_5": "全く迅速でない",
-    "csat_question_7_headline": "私たちのサービスに関するあなたの質問に、私たちはどのくらい迅速に対応してきましたか？",
-    "csat_question_7_subheader": "一つ選択してください：",
-    "csat_question_8_choice_1": "初めての購入です",
-    "csat_question_8_choice_2": "6ヶ月未満",
-    "csat_question_8_choice_3": "6ヶ月から1年",
-    "csat_question_8_choice_4": "1年から2年",
-    "csat_question_8_choice_5": "3年以上",
-    "csat_question_8_headline": "$[workspaceName]の顧客になってどのくらい経ちますか？",
-    "csat_question_8_subheader": "一つ選択してください：",
-    "csat_question_9_choice_1": "可能性が非常に高い",
-    "csat_question_9_choice_2": "可能性がとても高い",
-    "csat_question_9_choice_3": "可能性がやや高い",
-    "csat_question_9_choice_4": "可能性があまり高くない",
-    "csat_question_9_choice_5": "可能性が全くない",
-    "csat_question_9_headline": "当社の$[workspaceName]を再度購入する可能性はどのくらいありますか？",
-    "csat_question_9_subheader": "一つ選択してください：",
     "csat_survey_name": "$[workspaceName]顧客満足度調査",
     "csat_survey_question_1_headline": "$[workspaceName]の体験にどの程度満足していますか？",
     "csat_survey_question_1_lower_label": "非常に不満",
     "csat_survey_question_1_upper_label": "非常に満足",
     "csat_survey_question_2_headline": "素晴らしい！体験を改善するために何かできることはありますか？",
     "csat_survey_question_2_placeholder": "ここに回答を入力してください...",
-    "csat_survey_question_3_headline": "申し訳ありません！体験を改善するために何かできることはありますか？",
-    "csat_survey_question_3_placeholder": "ここに回答を入力してください...",
     "csat_upper_label": "非常に満足",
     "cta_description": "情報を表示し、特定の行動を促す",
     "custom_survey_description": "テンプレートを使わずにアンケートを作成する。",
@@ -915,12 +845,6 @@
     "custom_survey_question_1_headline": "何を知りたいですか？",
     "custom_survey_question_1_placeholder": "ここに回答を入力してください...",
     "customer_effort_score_description": "機能を使うのがどれだけ簡単か判断する。",
-    "customer_effort_score_name": "顧客努力スコア（CES）",
-    "customer_effort_score_question_1_headline": "$[workspaceName]を使うと、[目標を追加]することが簡単にできます",
-    "customer_effort_score_question_1_lower_label": "全くそう思わない",
-    "customer_effort_score_question_1_upper_label": "非常にそう思う",
-    "customer_effort_score_question_2_headline": "ありがとうございます！[ADD GOAL]をより簡単にするにはどうすればよいですか？",
-    "customer_effort_score_question_2_placeholder": "ここに回答を入力してください...",
     "date": "日付",
     "date_description": "日付の選択を求める",
     "default_ending_card_button_label": "あなた自身のアンケートを作成する",
@@ -1270,11 +1194,6 @@
     "nps": "ネットプロモータースコア（NPS）",
     "nps_description": "ネットプロモータースコアを測定する（0-10）",
     "nps_lower_label": "可能性が全くない",
-    "nps_name": "ネットプロモータースコア（NPS）",
-    "nps_question_1_headline": "$[workspaceName]を友人や同僚に勧める可能性はどの程度ですか？",
-    "nps_question_1_lower_label": "可能性が低い",
-    "nps_question_1_upper_label": "可能性が非常に高い",
-    "nps_question_2_headline": "その評価をつけた理由は何ですか？",
     "nps_survey_name": "NPSアンケート",
     "nps_survey_question_1_headline": "$[workspaceName]を友人や同僚に勧める可能性はどの程度ですか？",
     "nps_survey_question_1_lower_label": "可能性が全くない",
@@ -1481,10 +1400,6 @@
     "smileys_survey_question_1_headline": "$[workspaceName]はいかがですか？",
     "smileys_survey_question_1_lower_label": "良くない",
     "smileys_survey_question_1_upper_label": "非常に満足",
-    "smileys_survey_question_2_button_label": "レビューを書く",
-    "smileys_survey_question_2_headline": "お聞きできて嬉しいです 🙏 レビューを書いてください！",
-    "smileys_survey_question_2_html": "<p class=\"fb-editor-paragraph\" dir=\"ltr\"><span>これは私たちにとって大きな助けになります。</span></p>",
-    "smileys_survey_question_3_button_label": "送信",
     "smileys_survey_question_3_headline": "お聞きして申し訳ありません！私たちがもっとうまくできることは一つ何ですか？",
     "smileys_survey_question_3_placeholder": "ここに回答を入力してください...",
     "smileys_survey_question_3_subheader": "あなたの体験を改善するためにお手伝いください。",
@@ -1492,10 +1407,6 @@
     "star_rating_survey_question_1_headline": "$[workspaceName]はいかがですか？",
     "star_rating_survey_question_1_lower_label": "非常に不満",
     "star_rating_survey_question_1_upper_label": "非常に満足",
-    "star_rating_survey_question_2_button_label": "レビューを書く",
-    "star_rating_survey_question_2_headline": "お聞きできて嬉しいです 🙏 レビューを書いてください！",
-    "star_rating_survey_question_2_html": "<p class=\"fb-editor-paragraph\" dir=\"ltr\"><span>これは私たちにとって大きな助けになります。</span></p>",
-    "star_rating_survey_question_3_button_label": "送信",
     "star_rating_survey_question_3_headline": "お聞きして申し訳ありません！私たちがもっとうまくできることは一つ何ですか？",
     "star_rating_survey_question_3_placeholder": "ここに回答を入力してください...",
     "star_rating_survey_question_3_subheader": "あなたの体験を改善するためにお手伝いください。",
