+++ apps/web/locales/zh-Hans-CN.json
@@ -316,7 +316,6 @@
     "more_options": "更多选项",
     "move_down": "下移",
     "move_up": "上移",
-    "my_product": "我的产品",
     "name": "名称",
     "new_version_available": "Formbricks {version} 在 这里。立即 升级！",
     "next": "下一步",
@@ -833,81 +832,12 @@
     "csat": "客户满意度 (CSAT)",
     "csat_description": "测量客户满意度评分 (1-5)",
     "csat_lower_label": "非常不满意",
-    "csat_name": "客户满意度得分 ( CSAT )",
-    "csat_question_10_headline": "你还有其他的评论、问题或担忧吗？",
-    "csat_question_10_placeholder": "在 此 输入 您 的 答案...",
-    "csat_question_1_headline": "您向朋友或同事推荐这个 $[workspaceName] 的可能性有多大？",
-    "csat_question_1_lower_label": "不可能",
-    "csat_question_1_upper_label": "非常 可能",
-    "csat_question_2_choice_1": "有点 满意",
-    "csat_question_2_choice_2": "非常 满意",
-    "csat_question_2_choice_3": "既不 满意 也 不 不满意",
-    "csat_question_2_choice_4": "有点 不满意",
-    "csat_question_2_choice_5": "非常 不 满意",
-    "csat_question_2_headline": "总体而言，您对我们的 $[workspaceName] 满意还是不满意？",
-    "csat_question_2_subheader": "请选择一个：",
-    "csat_question_3_choice_1": "无效",
-    "csat_question_3_choice_10": "独特",
-    "csat_question_3_choice_2": "有用",
-    "csat_question_3_choice_3": "不实用",
-    "csat_question_3_choice_4": "价格 过高",
-    "csat_question_3_choice_5": "高质量",
-    "csat_question_3_choice_6": "可靠",
-    "csat_question_3_choice_7": "物有所值",
-    "csat_question_3_choice_8": "质量差",
-    "csat_question_3_choice_9": "不可靠",
-    "csat_question_3_headline": "您会用以下哪些词来描述我们的 $[workspaceName]？",
-    "csat_question_3_subheader": "请选择所有适用项：",
-    "csat_question_4_choice_1": "极好",
-    "csat_question_4_choice_2": "非常 好",
-    "csat_question_4_choice_3": "有点 好",
-    "csat_question_4_choice_4": "不太好",
-    "csat_question_4_choice_5": "完全 不好",
-    "csat_question_4_headline": "我们的 $[workspaceName] 在多大程度上满足了您的需求？",
-    "csat_question_4_subheader": "选择一项：",
-    "csat_question_5_choice_1": "非常高质量",
-    "csat_question_5_choice_2": "高质量",
-    "csat_question_5_choice_3": "低质量",
-    "csat_question_5_choice_4": "非常低质量",
-    "csat_question_5_choice_5": "不 高 不 低",
-    "csat_question_5_headline": "您如何评价 $[workspaceName] 的质量？",
-    "csat_question_5_subheader": "选择一项：",
-    "csat_question_6_choice_1": "优秀",
-    "csat_question_6_choice_2": "高于 平均水平",
-    "csat_question_6_choice_3": "平均",
-    "csat_question_6_choice_4": "低于 平均",
-    "csat_question_6_choice_5": "差",
-    "csat_question_6_headline": "您如何评价 $[workspaceName] 的性价比？",
-    "csat_question_6_subheader": "请选择 一个：",
-    "csat_question_7_choice_1": "特别 响应",
-    "csat_question_7_choice_2": "非常及时",
-    "csat_question_7_choice_3": "响应 轻微",
-    "csat_question_7_choice_4": "不太 响应",
-    "csat_question_7_choice_5": "一点都不 响应",
-    "csat_question_7_headline": "我们 对 您 关于 我们 服务 的 问题 响应 得 如何？",
-    "csat_question_7_subheader": "请选择一个：",
-    "csat_question_8_choice_1": "这是 我 的 第一次 购买",
-    "csat_question_8_choice_2": "不到 六个月",
-    "csat_question_8_choice_3": "六个月 到 一年",
-    "csat_question_8_choice_4": "1年 到 2年",
-    "csat_question_8_choice_5": "3年 或 更久",
-    "csat_question_8_headline": "您成为 $[workspaceName] 客户有多久了？",
-    "csat_question_8_subheader": "请选择 一个：",
-    "csat_question_9_choice_1": "极有可能",
-    "csat_question_9_choice_2": "非常 可能",
-    "csat_question_9_choice_3": "有点可能",
-    "csat_question_9_choice_4": "不太可能",
-    "csat_question_9_choice_5": "完全不可能",
-    "csat_question_9_headline": "您再次购买我们的 $[workspaceName] 的可能性有多大？",
-    "csat_question_9_subheader": "选择一项：",
     "csat_survey_name": "$[workspaceName] 客户满意度调查",
     "csat_survey_question_1_headline": "您对 $[workspaceName] 的体验满意度如何？",
     "csat_survey_question_1_lower_label": "极度不满意",
     "csat_survey_question_1_upper_label": "极度满意",
     "csat_survey_question_2_headline": "太好了！ 我们可以做些什么来改善您的体验？",
     "csat_survey_question_2_placeholder": "在此输入您的答案...",
-    "csat_survey_question_3_headline": "糟糕, 对不起！我们可以做些什么来改善您的体验？",
-    "csat_survey_question_3_placeholder": "在此输入您的答案...",
     "csat_upper_label": "非常满意",
     "cta_description": "显示 信息 并 提示用户采取 特定行动",
     "custom_survey_description": "创建 一个 没有 模板 的 调查。",
@@ -915,12 +845,6 @@
     "custom_survey_question_1_headline": "你 想 知道 什么？",
     "custom_survey_question_1_placeholder": "在此输入您的答案...",
     "customer_effort_score_description": "确定 使用 某个 功能 的 容易 程度。",
-    "customer_effort_score_name": "客户 努力 得分 (CES)",
-    "customer_effort_score_question_1_headline": "$[workspaceName] 让我能够轻松地[添加目标]",
-    "customer_effort_score_question_1_lower_label": "强烈 不同意",
-    "customer_effort_score_question_1_upper_label": "强烈 同意",
-    "customer_effort_score_question_2_headline": "谢谢！我们怎样才能让你更轻松地 [ADD GOAL] ?",
-    "customer_effort_score_question_2_placeholder": "在此输入您的答案...",
     "date": "日期",
     "date_description": "请求选择 日期",
     "default_ending_card_button_label": "创建 你 的 调查",
@@ -1270,11 +1194,6 @@
     "nps": "净推荐值 ( NPS )",
     "nps_description": "测量净推荐分数 ( 0-10 )",
     "nps_lower_label": "完全不可能",
-    "nps_name": "净推荐值 ( NPS )",
-    "nps_question_1_headline": "你向朋友或同事推荐 $[workspaceName] 的可能性有多大？",
-    "nps_question_1_lower_label": "不可能",
-    "nps_question_1_upper_label": "非常 可能",
-    "nps_question_2_headline": "是什么原因让你给出这个评分？",
     "nps_survey_name": "NPS 调查",
     "nps_survey_question_1_headline": "你向朋友或同事推荐 $[workspaceName] 的可能性有多大？",
     "nps_survey_question_1_lower_label": "完全不可能",
@@ -1481,10 +1400,6 @@
     "smileys_survey_question_1_headline": "您觉得$[workspaceName]怎么样？",
     "smileys_survey_question_1_lower_label": "不 好",
     "smileys_survey_question_1_upper_label": "非常 满意",
-    "smileys_survey_question_2_button_label": "写 评价",
-    "smileys_survey_question_2_headline": "很高兴听到 🙏 请为我们写一个评价！",
-    "smileys_survey_question_2_html": "<p class=\"fb-editor-paragraph\" dir=\"ltr\"><span>这 对 我们 帮助 很大。</span></p>",
-    "smileys_survey_question_3_button_label": "发送",
     "smileys_survey_question_3_headline": "很遗憾 听到！我们 可以 改进 的 一 件事 是 什么？",
     "smileys_survey_question_3_placeholder": "在 此 输入 您 的 答案...",
     "smileys_survey_question_3_subheader": "帮助 我们 改善 您 的 体验。",
@@ -1492,10 +1407,6 @@
     "star_rating_survey_question_1_headline": "您觉得$[workspaceName]怎么样？",
     "star_rating_survey_question_1_lower_label": "极度不满意",
     "star_rating_survey_question_1_upper_label": "极度满意",
-    "star_rating_survey_question_2_button_label": "写 评价",
-    "star_rating_survey_question_2_headline": "很高兴听到反馈 🙏 请为我们写一个评价！",
-    "star_rating_survey_question_2_html": "<p class=\"fb-editor-paragraph\" dir=\"ltr\"><span>这 对 我们 帮助 很大。</span></p>",
-    "star_rating_survey_question_3_button_label": "发送",
     "star_rating_survey_question_3_headline": "很遗憾 听到！我们 可以 改进 的 一 件事 是 什么？",
     "star_rating_survey_question_3_placeholder": "在此输入您的答案...",
     "star_rating_survey_question_3_subheader": "帮助 我们 改善 您 的 体验。",
