+++ apps/web/locales/zh-Hant-TW.json
@@ -316,7 +316,6 @@
     "more_options": "更多選項",
     "move_down": "下移",
     "move_up": "上移",
-    "my_product": "我的產品",
     "name": "名稱",
     "new_version_available": "Formbricks '{'version'}' 已推出。立即升級！",
     "next": "下一步",
@@ -833,81 +832,12 @@
     "csat": "顧客滿意度 (CSAT)",
     "csat_description": "測量顧客滿意度分數 (1-5)",
     "csat_lower_label": "非常不滿意",
-    "csat_name": "客戶滿意度分數 (CSAT)",
-    "csat_question_10_headline": "您有任何其他意見、問題或疑慮嗎？",
-    "csat_question_10_placeholder": "在此輸入您的答案...",
-    "csat_question_1_headline": "您向朋友或同事推薦這個 $[workspaceName] 的可能性有多高？",
-    "csat_question_1_lower_label": "不太可能",
-    "csat_question_1_upper_label": "非常可能",
-    "csat_question_2_choice_1": "有點滿意",
-    "csat_question_2_choice_2": "非常滿意",
-    "csat_question_2_choice_3": "既不滿意也不不滿意",
-    "csat_question_2_choice_4": "有點不滿意",
-    "csat_question_2_choice_5": "非常不滿意",
-    "csat_question_2_headline": "整體而言，您對我們的 $[workspaceName] 滿意或不滿意的程度如何？",
-    "csat_question_2_subheader": "請選取其中一項：",
-    "csat_question_3_choice_1": "無效",
-    "csat_question_3_choice_10": "獨特的",
-    "csat_question_3_choice_2": "有用的",
-    "csat_question_3_choice_3": "不切實際",
-    "csat_question_3_choice_4": "價格過高",
-    "csat_question_3_choice_5": "高品質",
-    "csat_question_3_choice_6": "可靠",
-    "csat_question_3_choice_7": "物有所值",
-    "csat_question_3_choice_8": "品質差",
-    "csat_question_3_choice_9": "不可靠",
-    "csat_question_3_headline": "您會用下列哪些詞來描述我們的 $[workspaceName]？",
-    "csat_question_3_subheader": "請選取所有適用的選項：",
-    "csat_question_4_choice_1": "非常好",
-    "csat_question_4_choice_2": "很好",
-    "csat_question_4_choice_3": "還可以",
-    "csat_question_4_choice_4": "不太好",
-    "csat_question_4_choice_5": "完全不好",
-    "csat_question_4_headline": "我們的 $[workspaceName] 符合您需求的程度如何？",
-    "csat_question_4_subheader": "選取一個選項：",
-    "csat_question_5_choice_1": "非常高品質",
-    "csat_question_5_choice_2": "高品質",
-    "csat_question_5_choice_3": "低品質",
-    "csat_question_5_choice_4": "非常低品質",
-    "csat_question_5_choice_5": "不高也不低",
-    "csat_question_5_headline": "您如何評價 $[workspaceName] 的品質？",
-    "csat_question_5_subheader": "選取一個選項：",
-    "csat_question_6_choice_1": "極佳",
-    "csat_question_6_choice_2": "高於平均",
-    "csat_question_6_choice_3": "平均",
-    "csat_question_6_choice_4": "低於平均",
-    "csat_question_6_choice_5": "差",
-    "csat_question_6_headline": "您如何評價 $[workspaceName] 的性價比？",
-    "csat_question_6_subheader": "請選取其中一項：",
-    "csat_question_7_choice_1": "非常快速回應",
-    "csat_question_7_choice_2": "非常快速回應",
-    "csat_question_7_choice_3": "有點快速回應",
-    "csat_question_7_choice_4": "不太快速回應",
-    "csat_question_7_choice_5": "完全不快速回應",
-    "csat_question_7_headline": "我們對您有關我們服務的問題的回應有多迅速？",
-    "csat_question_7_subheader": "請選取其中一項：",
-    "csat_question_8_choice_1": "這是我的第一次購買",
-    "csat_question_8_choice_2": "不到六個月",
-    "csat_question_8_choice_3": "六個月到一年",
-    "csat_question_8_choice_4": "1 - 2 年",
-    "csat_question_8_choice_5": "3 年或以上",
-    "csat_question_8_headline": "您成為 $[workspaceName] 客戶多久了？",
-    "csat_question_8_subheader": "請選取其中一項：",
-    "csat_question_9_choice_1": "非常有可能",
-    "csat_question_9_choice_2": "非常有可能",
-    "csat_question_9_choice_3": "有點可能",
-    "csat_question_9_choice_4": "不太可能",
-    "csat_question_9_choice_5": "完全不可能",
-    "csat_question_9_headline": "您再次購買我們任何 $[workspaceName] 的可能性有多高？",
-    "csat_question_9_subheader": "選取一個選項：",
     "csat_survey_name": "$[workspaceName] CSAT",
     "csat_survey_question_1_headline": "您對 $[workspaceName] 體驗的滿意度如何？",
     "csat_survey_question_1_lower_label": "極度不滿意",
     "csat_survey_question_1_upper_label": "極度滿意",
     "csat_survey_question_2_headline": "太棒了！我們是否有任何可以改善您體驗的地方？",
     "csat_survey_question_2_placeholder": "在此輸入您的答案...",
-    "csat_survey_question_3_headline": "唉，抱歉！我們是否有任何可以改善您體驗的地方？",
-    "csat_survey_question_3_placeholder": "在此輸入您的答案...",
     "csat_upper_label": "非常滿意",
     "cta_description": "顯示資訊並提示使用者採取特定操作",
     "custom_survey_description": "建立沒有範本的問卷。",
@@ -915,12 +845,6 @@
     "custom_survey_question_1_headline": "您想瞭解什麼？",
     "custom_survey_question_1_placeholder": "在此輸入您的答案...",
     "customer_effort_score_description": "判斷使用功能有多容易。",
-    "customer_effort_score_name": "客戶費力分數 (CES)",
-    "customer_effort_score_question_1_headline": "$[workspaceName] 讓我能輕鬆地 [ADD GOAL]",
-    "customer_effort_score_question_1_lower_label": "非常不同意",
-    "customer_effort_score_question_1_upper_label": "非常同意",
-    "customer_effort_score_question_2_headline": "謝謝！我們可以如何讓您更輕鬆地 [新增目標]？",
-    "customer_effort_score_question_2_placeholder": "在此輸入您的答案...",
     "date": "日期",
     "date_description": "要求選擇日期",
     "default_ending_card_button_label": "建立您自己的問卷",
@@ -1270,11 +1194,6 @@
     "nps": "淨推薦分數 (NPS)",
     "nps_description": "衡量淨推薦分數 (0-10)",
     "nps_lower_label": "完全不可能",
-    "nps_name": "淨推薦分數 (NPS)",
-    "nps_question_1_headline": "你有多大可能會向朋友或同事推薦 $[workspaceName]?",
-    "nps_question_1_lower_label": "不太可能",
-    "nps_question_1_upper_label": "非常有可能",
-    "nps_question_2_headline": "是什麼讓您給予此評分？",
     "nps_survey_name": "NPS 問卷",
     "nps_survey_question_1_headline": "你有多大可能會向朋友或同事推薦 $[workspaceName]?",
     "nps_survey_question_1_lower_label": "完全不可能",
@@ -1481,10 +1400,6 @@
     "smileys_survey_question_1_headline": "你覺得 $[workspaceName] 如何？",
     "smileys_survey_question_1_lower_label": "不好",
     "smileys_survey_question_1_upper_label": "非常滿意",
-    "smileys_survey_question_2_button_label": "撰寫評論",
-    "smileys_survey_question_2_headline": "很高興聽見 🙏 請為我們撰寫評論！",
-    "smileys_survey_question_2_html": "<p class=\"fb-editor-paragraph\" dir=\"ltr\"><span>這對我們有很大的幫助。</span></p>",
-    "smileys_survey_question_3_button_label": "發送",
     "smileys_survey_question_3_headline": "很抱歉聽到！我們應該改進哪一件事？",
     "smileys_survey_question_3_placeholder": "在此輸入您的答案...",
     "smileys_survey_question_3_subheader": "協助我們改善您的體驗。",
@@ -1492,10 +1407,6 @@
     "star_rating_survey_question_1_headline": "你覺得 $[workspaceName] 如何？",
     "star_rating_survey_question_1_lower_label": "極度不滿意",
     "star_rating_survey_question_1_upper_label": "極度滿意",
-    "star_rating_survey_question_2_button_label": "撰寫評論",
-    "star_rating_survey_question_2_headline": "很高興聽見 🙏 請為我們撰寫評論！",
-    "star_rating_survey_question_2_html": "<p class=\"fb-editor-paragraph\" dir=\"ltr\"><span>這對我們有很大的幫助。</span></p>",
-    "star_rating_survey_question_3_button_label": "發送",
     "star_rating_survey_question_3_headline": "很抱歉聽到！我們應該改進哪一件事？",
     "star_rating_survey_question_3_placeholder": "在此輸入您的答案...",
     "star_rating_survey_question_3_subheader": "協助我們改善您的體驗。",
